Update: Refactor code

+++ src/External/Presentation/PenomyAPI.Presentation.FastEndpointBasedApi/Endpoints/FeatG5/DTOs/G5ResponseDto.cs
@@ -13,7 +13,7 @@ namespace PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG5.DTOs
         public string SeriesName { get; set; }
         public long ViewCount { get; set; }
         public long FavoriteCount { get; set; }
-        public byte StarRates { get; set; }
+        public double StarRates { get; set; }
         public string ThumbnailUrl { get; set; }
     }
 }
